backend/requirements.txt
Django
djangorestframework
scikit-learn
spacy
